// apps/web/src/api-client/model/adminReportStatusHistoryItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */
import type { AdminReportStatusHistoryItemOldStatus } from "./adminReportStatusHistoryItemOldStatus";
import type { ReportStatus } from "./reportStatus";
import type { AdminReportStatusHistoryItemChangedBy } from "./adminReportStatusHistoryItemChangedBy";

export interface AdminReportStatusHistoryItem {
    /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
    id: string;
    /** @nullable */
    oldStatus: AdminReportStatusHistoryItemOldStatus;
    newStatus: ReportStatus;
    /** @nullable */
    reason: string | null;
    createdAt: string;
    /** @nullable */
    changedBy: AdminReportStatusHistoryItemChangedBy;
}